tgbot-verify/requirements.txt
python-telegram-bot>=20.0
httpx>=0.27.0
Pillow>=10.0.0
reportlab>=4.0.0
xhtml2pdf>=0.2.15
playwright==1.48.0
pymysql>=1.1.0
psutil>=5.9.0
python-dotenv>=1.0.0
